brain_tumor_project/requirements.txt
numpy
matplotlib
pillow
scikit-learn
tensorflow>=2.9
opencv-python
streamlit
